Refactor theme creation
